+++ crypto/encrypt_test.go
@@ -9,15 +9,16 @@ import (
 
 func Test_Encrypt_Decrypt(t *testing.T) {
 	t.Parallel()
+	ctx := t.Context()
 
-	key, err := CreateKey(32)
+	key, err := CreateKey(ctx, 32)
 	require.NoError(t, err)
 	plaintext := "Mary Jane Hawkins"
 
 	ciphertext, err := Encrypt(key, []byte(plaintext))
 	require.NoError(t, err)
 
-	decryptedPlaintext, err := Decrypt(key, ciphertext)
+	decryptedPlaintext, err := Decrypt(ctx, key, ciphertext)
 	require.NoError(t, err)
 	assert.Equal(t, plaintext, string(decryptedPlaintext))
 }