+++ gomock_generator/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## To be Released
 
 * feat(drop-gopath) Stop basing all generation on GOPATH which is not mandatory anymore
+* refactor: only use `github.com/Scalingo/go-utils/errors/v3` for errors
 
 ## v1.4.2